// src/app/models/parcela.model.ts
export class ParcelaModel {

  id: string;
  activa: boolean;
  numeroParcela: number;
  fechaIngreso: Date;
  categoria: string;
  cantidad: number;
  dueño: string;
  proximaParcela: number;
  agua: boolean;
  sombra: boolean;
  estado: string;
  remanente: string;
  malezas: [];
  desmalezar: boolean;
  ultimaLluvia: Date;
  cantidadLluvia: number;
  hongos: boolean;
  hormigueros: boolean;

}